backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# MongoDB async driver
motor==3.6.0
pymongo==4.9.2

# Pydantic v2 + settings (aiogram 3.13 requires pydantic<2.10)
pydantic[email]==2.9.2
pydantic-settings==2.6.1

# JWT + security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.3

# Telegram bot
aiogram==3.13.1

# OpenAI
openai==1.57.2

# OCR
pytesseract==0.3.13
Pillow==11.0.0

# HTTP client (for webhooks)
httpx==0.27.2

# Scheduler (archive tasks)
APScheduler==3.10.4

# Redis (optional: session/cache)
redis==5.2.1

# File upload
python-multipart==0.0.19

# Environment
python-dotenv==1.0.1
